// app/utils/enum.ts
export type NullableDateString = string | null;
export type NullableNumber = number | null;

export interface SinanTuberculoseRecord {
  TP_NOT: string;
  ID_AGRAVO: string;
  DT_NOTIFIC: NullableDateString;
  NU_ANO: string;
  SG_UF_NOT: string;
  ID_MUNICIP: string;
  ID_REGIONA: string;
  DT_DIAG: NullableDateString;
  ANO_NASC: string;
  NU_IDADE_N: NullableNumber;
  CS_SEXO: string;
  CS_GESTANT: string;
  CS_RACA: string;
  CS_ESCOL_N: string;
  SG_UF: string;
  ID_MN_RESI: string;
  ID_RG_RESI: string;
  ID_PAIS: string;
  NDUPLIC_N: string;
  IN_VINCULA: string;
  DT_DIGITA: NullableDateString;
  DT_TRANSUS: NullableDateString;
  DT_TRANSDM: NullableDateString;
  DT_TRANSSM: NullableDateString;
  DT_TRANSRM: NullableDateString;
  DT_TRANSRS: NullableDateString;
  DT_TRANSSE: NullableDateString;
  CS_FLXRET: string;
  FLXRECEBI: string;
  MIGRADO_W: string;
  ID_OCUPA_N: string;
  TRATAMENTO: string;
  INSTITUCIO: string;
  RAIOX_TORA: string;
  TESTE_TUBE: string;
  FORMA: string;
  EXTRAPU1_N: string;
  EXTRAPU2_N: string;
  AGRAVAIDS: string;
  AGRAVALCOO: string;
  AGRAVDIABE: string;
  AGRAVDOENC: string;
  AGRAVOUTRA: string;
  BACILOSC_E: string;
  BACILOS_E2: string;
  BACILOSC_O: string;
  CULTURA_ES: string;
  CULTURA_OU: string;
  HIV: string;
  HISTOPATOL: string;
  DT_INIC_TR: NullableDateString;
  RIFAMPICIN: string;
  ISONIAZIDA: string;
  ETAMBUTOL: string;
  ESTREPTOMI: string;
  PIRAZINAMI: string;
  ETIONAMIDA: string;
  OUTRAS: string;
  TRAT_SUPER: string;
  NU_CONTATO: NullableNumber;
  DOENCA_TRA: string;
  SG_UF_AT: string;
  ID_MUNIC_A: string;
  DT_NOTI_AT: NullableDateString;
  SG_UF_2: string;
  ID_MUNIC_2: string;
  BACILOSC_1: string;
  BACILOSC_2: string;
  BACILOSC_3: string;
  BACILOSC_4: string;
  BACILOSC_5: string;
  BACILOSC_6: string;
  TRATSUP_AT: string;
  DT_MUDANCA: NullableDateString;
  NU_COMU_EX: NullableNumber;
  SITUA_9_M: string;
  SITUA_12_M: string;
  SITUA_ENCE: string;
  DT_ENCERRA: NullableDateString;
  TPUNINOT: string;
  POP_LIBER: string;
  POP_RUA: string;
  POP_SAUDE: string;
  POP_IMIG: string;
  BENEF_GOV: string;
  AGRAVDROGA: string;
  AGRAVTABAC: string;
  TEST_MOLEC: string;
  TEST_SENSI: string;
  ANT_RETRO: string;
  BAC_APOS_6: string;
  TRANSF: string;
  UF_TRANSF: string;
  MUN_TRANSF: string;
}

export interface Data {
    notifications: SinanTuberculoseRecord[]
}